Created auth route and initialized React project

+++ z2/controllers/Todos.js
@@ -2,8 +2,10 @@ const db = require("../config/db");
 const Todo = db.todos;
 
 const getTodos = async (req, res) => {
+  if (!req.isAuth) return res.json({ message: "Access Denied" });
+  const userId = req.id;
   try {
-    const todos = await Todo.findAll();
+    const todos = await Todo.findAll({ where: { userId } });
     return res.status(200).json(todos);
   } catch (err) {
     console.log(err);
@@ -11,18 +13,22 @@ const getTodos = async (req, res) => {
   }
 };
 
-const getTodo = async (req, res) => {
-  const id = req.params.id;
-  try {
-    const todo = await Todo.findByPk(id);
-    return res.status(200).json(todo);
-  } catch (err) {
-    console.log(err);
-    return res.status(400).json(err);
-  }
-};
+// const getTodo = async (req, res) => {
+//   if (!req.isAuth) return res.json({message: 'Access Denied'})
+//   const userId = req.id
+//   const id = req.params.id;
+//   try {
+//     const todo = await Todo.findByPk(id);
+//     return res.status(200).json(todo);
+//   } catch (err) {
+//     console.log(err);
+//     return res.status(400).json(err);
+//   }
+// };
 
 const createTodo = async (req, res) => {
+  if (!req.isAuth) return res.json({ message: "Access Denied" });
+  const userId = req.id;
   const { title, checked, deadline } = req.body;
   console.log(Todo);
   try {
@@ -30,6 +36,7 @@ const createTodo = async (req, res) => {
       title,
       checked,
       deadline,
+      userId,
     });
     return res.status(200).json(todo);
   } catch (err) {
@@ -39,12 +46,14 @@ const createTodo = async (req, res) => {
 };
 
 const updateTodo = async (req, res) => {
+  if (!req.isAuth) return res.json({ message: "Access Denied" });
+  const userId = req.id;
   const id = req.params.id;
   const { title, checked, deadline } = req.body;
   try {
     const todo = await Todo.update(
       { title, checked, deadline },
-      { where: { id } }
+      { where: { id, userId } }
     );
     return res.status(200).json(todo);
   } catch (err) {
@@ -54,9 +63,11 @@ const updateTodo = async (req, res) => {
 };
 
 const deleteTodo = async (req, res) => {
+  if (!req.isAuth) return res.json({ message: "Access Denied" });
+  const userId = req.id;
   const id = req.params.id;
   try {
-    const todo = await Todo.destroy({ where: { id } });
+    const todo = await Todo.destroy({ where: { id, userId } });
     return res.status(200).json(todo);
   } catch (err) {
     console.log(err);
@@ -66,7 +77,7 @@ const deleteTodo = async (req, res) => {
 
 module.exports = {
   getTodos,
-  getTodo,
+  // getTodo,
   createTodo,
   updateTodo,
   deleteTodo,
